Add files via upload
https://leetcode.com/problems/frog-jump/
A frog needs to cross a river divided into units with or without stones. The frog can jump on stones, but not in the water. The frog starts on the first stone, and its jumps must be 1 unit initially, and then either one less, the same, or one more unit as the previous jump, only in the forward direction. This code determines whether, for a given sorted list of stone positions, the frog can successfully reach the last stone.